+++ r/tests/testthat/helper-taco.R
@@ -1,1 +1,2 @@
 taco_fixture <- function() testthat::test_path("data", "taco.zip")
+taco_nested_fixture <- function() testthat::test_path("data", "taco_nested.zip")